+++ src/pages/rest-area-fuel/RestAreaFuelPage.tsx
@@ -1,12 +1,5 @@
-import { AsyncBoundary } from '#/components/async-boundary';
-import {
-    RestAreaFuelInformation,
-    RestAreaFuelInformationLoading,
-} from '#/features/rest-area/rest-area-fuel-information';
-import {
-    RestAreaParkingInformation,
-    RestAreaParkingInformationLoading,
-} from '#/features/rest-area/rest-area-parking-information';
+import { RestAreaFuelInformation } from '#/features/rest-area/rest-area-fuel-information';
+import { RestAreaParkingInformation } from '#/features/rest-area/rest-area-parking-information';
 import {
     useGetRestAreaGasStationInfo,
     useGetRestAreaParkingInfo,
@@ -20,26 +13,20 @@ export const RestAreaFuelPage = () => {
 
     return (
         <S.Container gap={8}>
-            <AsyncBoundary pendingFallback={<RestAreaFuelInformationLoading />}>
-                <RestAreaFuelInformation
-                    gasolinePrice={fuel.gasolinePrice}
-                    dieselPrice={fuel.dieselPrice}
-                    lpgPrice={fuel.lpgPrice}
-                    isElectricChargingStation={fuel.isElectricChargingStation}
-                    isHydrogenChargingStation={fuel.isHydrogenChargingStation}
-                />
-            </AsyncBoundary>
-            <AsyncBoundary
-                pendingFallback={<RestAreaParkingInformationLoading />}
-            >
-                <RestAreaParkingInformation
-                    smallCarSpace={parking.smallCarSpace}
-                    largeCarSpace={parking.largeCarSpace}
-                    disabledPersonSpace={parking.disabledPersonSpace}
-                    totalSpace={parking.totalSpace}
-                    updateDate={parking.updateDate}
-                />
-            </AsyncBoundary>
+            <RestAreaFuelInformation
+                gasolinePrice={fuel.gasolinePrice}
+                dieselPrice={fuel.dieselPrice}
+                lpgPrice={fuel.lpgPrice}
+                isElectricChargingStation={fuel.isElectricChargingStation}
+                isHydrogenChargingStation={fuel.isHydrogenChargingStation}
+            />
+            <RestAreaParkingInformation
+                smallCarSpace={parking.smallCarSpace}
+                largeCarSpace={parking.largeCarSpace}
+                disabledPersonSpace={parking.disabledPersonSpace}
+                totalSpace={parking.totalSpace}
+                updateDate={parking.updateDate}
+            />
         </S.Container>
     );
 };
