+++ src/pages/rest-area-fuel/index.ts
@@ -1,1 +1,4 @@
-export { RestAreaFuelPage } from './RestAreaFuelPage';
+import { RestAreaFuelPage } from './RestAreaFuelPage';
+import { RestAreaFuelPageLoading } from './RestAreaFuelPage.loading';
+
+export { RestAreaFuelPage, RestAreaFuelPageLoading };
