+++ src/pages/templates/rest-area-detail/index.ts
@@ -1,4 +1,11 @@
-export {
+import {
     RestAreaDetail,
     type RestAreaDetailOutletContextType,
 } from './RestAreaDetail';
+import { RestAreaDetailLoading } from './RestAreaDetail.loading';
+
+export {
+    RestAreaDetailLoading,
+    RestAreaDetail,
+    type RestAreaDetailOutletContextType,
+};
