+++ src/routers/index.tsx
@@ -1,3 +1,5 @@
+import { Suspense } from 'react';
+
 import { ThemeProvider } from '@emotion/react';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
@@ -8,15 +10,24 @@ import { AsyncBoundary } from '#/components/async-boundary';
 import { InternalErrorPage } from '#/pages/internal-error';
 import { LocationSearch } from '#/pages/location-search';
 import { NotFoundPage } from '#/pages/not-found';
-import { RestAreaFoodPage } from '#/pages/rest-area-food';
-import { RestAreaFuelPage } from '#/pages/rest-area-fuel';
-import { RestAreaMapPage } from '#/pages/rest-area-map';
+import {
+    RestAreaFoodPage,
+    RestAreaFoodPageLoading,
+} from '#/pages/rest-area-food';
+import {
+    RestAreaFuelPage,
+    RestAreaFuelPageLoading,
+} from '#/pages/rest-area-fuel';
+import { RestAreaMapPage, RestAreaMapPageLoading } from '#/pages/rest-area-map';
 import {
     RestAreaOtherInformation,
     RestAreaOtherInformationLoading,
 } from '#/pages/rest-area-other-information';
 import { MobileView } from '#/pages/templates/mobile-view';
-import { RestAreaDetail } from '#/pages/templates/rest-area-detail';
+import {
+    RestAreaDetail,
+    RestAreaDetailLoading,
+} from '#/pages/templates/rest-area-detail';
 import { GlobalStyle } from '#/styles/global';
 import { theme } from '#/styles/theme';
 
@@ -58,15 +69,31 @@ export const applicationRouter: ReturnType<typeof createBrowserRouter> =
                 {
                     path: '/rest-area/:restAreaId',
                     errorElement: <InternalErrorPage />,
-                    element: <RestAreaDetail />,
+                    element: (
+                        <Suspense fallback={<RestAreaDetailLoading />}>
+                            <RestAreaDetail />
+                        </Suspense>
+                    ),
                     children: [
                         {
                             path: 'foods',
-                            element: <RestAreaFoodPage />,
+                            element: (
+                                <Suspense
+                                    fallback={<RestAreaFoodPageLoading />}
+                                >
+                                    <RestAreaFoodPage />
+                                </Suspense>
+                            ),
                         },
                         {
                             path: 'fuel-parking',
-                            element: <RestAreaFuelPage />,
+                            element: (
+                                <Suspense
+                                    fallback={<RestAreaFuelPageLoading />}
+                                >
+                                    <RestAreaFuelPage />
+                                </Suspense>
+                            ),
                         },
                         {
                             path: 'other-information',
@@ -85,7 +112,11 @@ export const applicationRouter: ReturnType<typeof createBrowserRouter> =
                 {
                     path: '/map',
                     errorElement: <InternalErrorPage />,
-                    element: <RestAreaMapPage />,
+                    element: (
+                        <Suspense fallback={<RestAreaMapPageLoading />}>
+                            <RestAreaMapPage />
+                        </Suspense>
+                    ),
                 },
             ],
             errorElement: (
